// packages/shared/src/types/collection.types.d.ts
export interface ApiEndpointHeader {
    key: string;
    value: string;
    enabled: boolean;
}
/**
 * API Endpoint 정의
 */
export interface ApiEndpointDefinition {
    id: string;
    name: string;
    baseUrl: string;
    path: string;
    method?: string;
    headers?: ApiEndpointHeader[] | Record<string, string>;
}